// src/types/braveSearch.ts
export type FormattedWebResults = {
  url: string;
  title: string;
  description?: string;
  extra_snippets?: string[];
}[];

export type FormattedFAQResults = {
  question: string;
  answer: string;
  title: string;
  url: string;
}[];

export type FormattedDiscussionsResults = {
  mutated_by_goggles?: boolean;
  url: string;
  data?: ForumData;
}[];

export type FormattedNewsResults = {
  mutated_by_goggles?: boolean;
  source?: string;
  breaking: boolean;
  is_live: boolean;
  age?: string;
  url: string;
  title: string;
  description?: string;
  extra_snippets?: string[];
}[];

export type FormattedVideoResults = {
  mutated_by_goggles?: boolean;
  url: string;
  title: string;
  description?: string;
  age?: string;
  thumbnail_url?: string;
  duration?: string;
  view_count?: string;
  creator?: string;
  publisher?: string;
  tags?: string[];
}[];
// ...
export interface Discussions {
  /** The type identifying a discussion cluster. Currently the value is always search. */
  type: 'search';
  /** A list of discussion results. */
  results: DiscussionResult[];
  /** Whether the discussion results are changed by a Goggle. The value is false by default. */
  mutated_by_goggles: boolean;
}

interface DiscussionResult extends Omit<SearchResult, 'type'> {
  /** The discussion result type identifier. The value is always discussion. */
  type: 'discussion';
  /** The enriched aggregated data for the relevant forum post. */
  data?: ForumData;
}

export interface ForumData {
  /** The name of the forum. */
  forum_name: string;
  /** The number of answers to the post. */
  num_answers?: number;
  /** The score of the post on the forum. */
  score?: string;
  /** The title of the post on the forum. */
  title?: string;
  /** The question asked in the forum post. */
  question?: string;
  /** The top-rated comment under the forum post. */
  top_comment?: string;
}

export interface FAQ {
  /** The FAQ result type identifier. The value is always faq. */
  type: 'faq';
  /** A list of aggregated question answer results relevant to the query. */
  results: QA[];
}

interface QA {
  /** The question being asked. */
  question: string;
  /** The answer to the question. */
  answer: string;
  /** The title of the post. */
  title: string;
  /** The URL pointing to the post. */
  url: string;
  /** Aggregated information about the URL. */
  meta_url?: MetaUrl;
}

interface MetaUrl {
  /** The protocol scheme extracted from the URL. */
  scheme: string;
  /** The network location part extracted from the URL. */
  netloc: string;
  /** The lowercased domain name extracted from the URL. */
  hostname?: string;
  /** The favicon used for the URL. */
  favicon: string;
  /** The hierarchical path of the URL useful as a display string. */
  path: string;
}

export interface Search {
  /** A type identifying web search results. The value is always search. */
  type: 'search';
  /** A list of search results. */
  results: SearchResult[];
  /** Whether the results are family friendly. */
  family_friendly: boolean;
}

export interface SearchResult extends Result {
  /** A type identifying a web search result. The value is always search_result. */
  type: 'search_result';
  /** A sub type identifying the web search result type. */
  subtype: 'generic';
  /** Whether the web search result is currently live. Default value is false. */
  is_live: boolean;
  /** Gathered information on a web search result. */
  deep_results?: DeepResult;
  /** A list of schemas (structured data) extracted from the page. The schemas try to follow schema.org and will return anything we can extract from the HTML that can fit into these models. */
  schemas?: [][];
  /** Aggregated information on the URL associated with the web search result. */
  meta_url?: MetaUrl;
  /** The thumbnail of the web search result. */
  thumbnail?: Thumbnail;
  /** A string representing the age of the web search result. */
  age?: string;
  /** The main language on the web search result. */
  language: string;
  /** The location details if the query relates to a restaurant. */
  location?: LocationResult;
  /** The video associated with the web search result. */
  video?: VideoData;
  /** The movie associated with the web search result. */
  movie?: MovieData;
  /** Any frequently asked questions associated with the web search result. */
  faq?: FAQ;
  /** Any question answer information associated with the web search result page. */
  qa?: QAPage;
  /** Any book information associated with the web search result page. */
  book?: Book;
  /** Rating found for the web search result page. */
  rating?: Rating;
  /** An article found for the web search result page. */
  article?: Article;
  /** The main product and a review that is found on the web search result page. */
  product?: Product | Review;
  /** A list of products and reviews that are found on the web search result page. */
  product_cluster?: Product | Review[];
  /** A type representing a cluster. The value can be product_cluster. */
  cluster_type?: string;
  /** A list of web search results. */
  cluster?: Result[];
  /** Aggregated information on the creative work found on the web search result. */
  creative_work?: CreativeWork;
  /** Aggregated information on music recording found on the web search result. */
  music_recording?: MusicRecording;
  /** Aggregated information on the review found on the web search result. */
  review?: Review;
  /** Aggregated information on a software product found on the web search result page. */
  software?: Software;
  /** Aggregated information on a recipe found on the web search result page. */
  recipe?: Recipe;
  /** Aggregated information on a organization found on the web search result page. */
  organization?: Organization;
  /** The content type associated with the search result page. */
  content_type?: string;
  /** A list of extra alternate snippets for the web search result. */
  extra_snippets?: string[];
}

interface Result {
  /** The title of the web page. */
  title: string;
  /** The URL where the page is served. */
  url: string;
  is_source_local: boolean;
  is_source_both: boolean;
  /** A description for the web page. */
  description?: string;
  /** A date representing the age of the web page. */
  page_age?: string;
  /** A date representing when the web page was last fetched. */
  page_fetched?: string;
  /** A profile associated with the web page. */
  profile?: Profile;
  /** A language classification for the web page. */
  language?: string;
  /** Whether the web page is family friendly. */
  family_friendly: boolean;
}
// ...
interface Thumbnail {
  /** The served URL of the picture thumbnail. */
  src: string;
  /** The original URL of the image. */
  original?: string;
}

interface LocationWebResult extends Result {
  /** Aggregated information about the URL. */
  meta_url: MetaUrl;
}

interface LocationResult extends Result {
  /** Location result type identifier. The value is always location_result. */
  type: 'location_result';
  /** A Temporary id associated with this result, which can be used to retrieve extra information about the location. It remains valid for 8 hours… */
  id?: string;
  /** The complete URL of the provider. */
  provider_url: string;
  /** A list of coordinates associated with the location. This is a lat long represented as a floating point. */
  coordinates?: number[];
  /** The zoom level on the map. */
  zoom_level: number;
  /** The thumbnail associated with the location. */
  thumbnail?: Thumbnail;
  /** The postal address associated with the location. */
  postal_address?: PostalAddress;
  /** The opening hours, if it is a business, associated with the location . */
  opening_hours?: OpeningHours;
  /** The contact of the business associated with the location. */
  contact?: Contact;
  /** A display string used to show the price classification for the business. */
  price_range?: string;
  /** The ratings of the business. */
  rating?: Rating;
  /** The distance of the location from the client. */
  distance?: Unit;
  /** Profiles associated with the business. */
  profiles?: DataProvider[];
  /** Aggregated reviews from various sources relevant to the business. */
  reviews?: Reviews;
  /** A bunch of pictures associated with the business. */
  pictures?: PictureResults;
  /** An action to be taken. */
  action?: Action;
  /** A list of cuisine categories served. */
  serves_cuisine?: string[];
  /** A list of categories. */
  categories?: string[];
  /** An icon category. */
  icon_category?: string;
  /** Web results related to this location. */
  results?: LocationWebResult;
  /** IANA timezone identifier. */
  timezone?: string;
  /** The utc offset of the timezone. */
  timezone_offset?: string;
}
// ...
export interface Videos {
  /** The type representing the videos. The value is always videos. */
  type: 'videos';
  /** A list of video results. */
  results: VideoResult[];
  /** Whether the video results are changed by a Goggle. The value is false by default. */
  mutated_by_goggles?: boolean;
}

export interface News {
  /** The type representing the news. The value is always news. */
  type: 'news';
  /** A list of news results. */
  results: NewsResult[];
  /** Whether the news results are changed by a Goggle. The value is false by default. */
  mutated_by_goggles?: boolean;
}

interface NewsResult extends Result {
  /** The aggregated information on the URL representing a news result. */
  meta_url?: MetaUrl;
  /** The source of the news. */
  source?: string;
  /** Whether the news result is currently a breaking news. */
  breaking: boolean;
  /** Whether the news result is currently live. */
  is_live: boolean;
  /** The thumbnail associated with the news result. */
  thumbnail?: Thumbnail;
  /** A string representing the age of the news article. */
  age?: string;
  /** A list of extra alternate snippets for the news search result. */
  extra_snippets?: string[];
}
// ...
interface MovieData {
  /** Name of the movie. */
  name?: string;
  /** A short plot summary for the movie. */
  description?: string;
  /** A URL serving a movie profile page. */
  url?: string;
  /** A thumbnail for a movie poster. */
  thumbnail?: Thumbnail;
  /** The release date for the movie. */
  release?: string;
  /** A list of people responsible for directing the movie. */
  directors?: Person[];
  /** A list of actors in the movie. */
  actors?: Person[];
  /** Rating provided to the movie from various sources. */
  rating?: Rating;
  /** The runtime of the movie. The format is HH:MM:SS. */
  duration?: string;
  /** List of genres in which the movie can be classified. */
  genre?: string[];
  /** The query that resulted in the movie result. */
  query?: string;
}
// ...
interface DeepResult {
  /** A list of news results associated with the result. */
  news?: NewsResult[];
  /** A list of buttoned results associated with the result. */
  buttons?: ButtonResult[];
  /** Videos associated with the result. */
  videos?: VideoResult[];
  /** Images associated with the result. */
  images?: Image[];
}

interface VideoResult extends Result {
  /** The type identifying the video result. The value is always video_result. */
  type: 'video_result';
  /** Meta data for the video. */
  video: VideoData;
  /** Aggregated information on the URL. */
  meta_url?: MetaUrl;
  /** The thumbnail of the video. */
  thumbnail?: Thumbnail;
  /** A string representing the age of the video. */
  age?: string;
}

interface VideoData {
  /** A time string representing the duration of the video. The format can be HH:MM:SS or MM:SS. */
  duration?: string;
  /** The number of views of the video. */
  views?: string;
  /** The creator of the video. */
  creator?: string;
  /** The publisher of the video. */
  publisher?: string;
  /** A thumbnail associated with the video. */
  thumbnail?: Thumbnail;
  /** A list of tags associated with the video. */
  tags?: string[];
  /** Author of the video. */
  author?: Profile;
  /** Whether the video requires a subscription to watch. */
  requires_subscription?: boolean;
}